requirements.txt
fastapi==0.70.0
uvicorn==0.15.0
python-dotenv==0.19.1
